+++ DB/src/main/java/ies/jandula/universidad/repository/DepartamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ies.jandula.universidad.models.Departamento;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long>{
+public interface DepartamentoRepository extends JpaRepository<Departamento, Integer>{
 
 }
