+++ DB/src/main/java/ies/jandula/universidad/repository/GradoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ies.jandula.universidad.models.Grado;
 
-public interface GradoRepository extends JpaRepository<Grado, Long>{
+public interface GradoRepository extends JpaRepository<Grado, Integer>{
 
 }
