+++ DB/src/main/java/ies/jandula/universidad/repository/ProfesorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ies.jandula.universidad.models.Profesor;
 
-public interface ProfesorRepository extends JpaRepository<Profesor, Long>{
+public interface ProfesorRepository extends JpaRepository<Profesor, Integer>{
 
 }
